updating tmy landing
