+++ components/homepage/components/AutoSlider/Slideshow.css
@@ -1,0 +1,16 @@
+.slideshow {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 80%;
+  }
+  
+  .slide {
+    margin-bottom: 1rem;
+  }
+  
+  .controls {
+    display: flex;
+    gap: 1rem;
+    justify-content: center;
+  }
