+++ components/homepage/components/AutoSlider/Slideshow.tsx
@@ -1,0 +1,53 @@
+"use client"
+import { useEffect, useState } from 'react';
+import "./Slideshow.css"
+
+interface SlideshowProps {
+  images: string[];
+}
+
+const Slideshow: React.FC<SlideshowProps> = ({ images }) => {
+  const [currentImageIndex, setCurrentImageIndex] = useState(0);
+  const [isPlaying, setIsPlaying] = useState(true);
+
+  useEffect(() => {
+    let interval: NodeJS.Timeout;
+
+    if (isPlaying) {
+      interval = setInterval(() => {
+        setCurrentImageIndex((prevIndex) => (prevIndex + 1) % images.length);
+      }, 5000);
+    }
+
+    return () => {
+      clearInterval(interval);
+    };
+  }, [isPlaying]);
+
+  const handlePrev = () => {
+    setCurrentImageIndex((prevIndex) => (prevIndex - 1 + images.length) % images.length);
+  };
+
+  const handleNext = () => {
+    setCurrentImageIndex((prevIndex) => (prevIndex + 1) % images.length);
+  };
+
+  const handleTogglePlay = () => {
+    setIsPlaying((prevIsPlaying) => !prevIsPlaying);
+  };
+
+  return (
+    <div className="slideshow">
+      <div className="slide">
+        <img src={images[currentImageIndex]} alt="Slideshow Image" />
+      </div>
+      <div className="controls">
+        <button onClick={handlePrev}>Previous</button>
+        <button onClick={handleNext}>Next</button>
+        <button onClick={handleTogglePlay}>{isPlaying ? 'Stop' : 'Play'}</button>
+      </div>
+    </div>
+  );
+};
+
+export default Slideshow;
